Add time buffered logger factory functions
